+++ udata/core/reports/api.py
@@ -21,13 +21,13 @@ class ReportsAPI(API):
 
         return Report.apply_sort_filters_and_pagination(query)
 
-    @api.secure
     @api.doc("create_report", responses={400: "Validation error"})
     @api.expect(Report.__write_fields__)
     @api.marshal_with(Report.__read_fields__, code=201)
     def post(self):
         report = patch(Report(), request)
-        report.by = current_user._get_current_object()
+        if current_user.is_authenticated:
+            report.by = current_user._get_current_object()
 
         try:
             report.save()